+++ MMM-NHL.js
@@ -24,11 +24,13 @@ Module.register('MMM-NHL', {
         '1st': '1ST_PERIOD',
         '2nd': '2ND_PERIOD',
         '3rd': '3RD_PERIOD',
-        OT: 'OVER_TIME',
-        SO: 'SHOOTOUT',
-        FINAL: 'FINAL',
+        'OT': 'OVER_TIME',
+        'SO': 'SHOOTOUT',
+        'SHOOTOUT': 'SHOOTOUT',
+        'FINAL': 'FINAL',
         'FINAL OT': 'FINAL_OVERTIME',
-        'FINAL SO': 'FINAL_SHOOTOUT'
+        'FINAL SO': 'FINAL_SHOOTOUT',
+        'PPD': 'PPD'
     },
 
     teams: {
@@ -71,10 +73,14 @@ Module.register('MMM-NHL', {
     defaults: {
         colored: false,
         focus_on: false,
+        condensed: false,
         matches: 6,
         format: 'ddd h:mm',
         rotateInterval: 20 * 1000, // every 20 seconds
-        reloadInterval: 30 * 60 * 1000 // every 30 minutes
+        reloadInterval: 30 * 60 * 1000, // every 30 minutes
+        gameTimeReloadInterval: 60 * 1000, // every 60 seconds
+        showLogos: true,
+        showNames: true
     },
 
     getTranslations() {
@@ -132,7 +138,7 @@ Module.register('MMM-NHL', {
 
         if (!this.scores) {
             const text = document.createElement('div');
-            text.innerHTML = this.translate('LOADING');
+            text.innerHTML = this.nhlTranslate('LOADING');
             text.classList.add('dimmed', 'light');
             scores.appendChild(text);
         } else {
@@ -154,6 +160,16 @@ Module.register('MMM-NHL', {
         return wrapper;
     },
 
+    nhlTranslate(key) {
+        if (this.config.condensed) {
+            const newKey = `${key}-Condensed`;
+            const result = this.translate(newKey, undefined);
+            if (result !== newKey)
+                return result;
+        }
+        return this.translate(key);
+    },
+
     createLabelRow() {
         const labelRow = document.createElement('tr');
 
@@ -164,7 +180,7 @@ Module.register('MMM-NHL', {
         labelRow.appendChild(dateLabel);
 
         const homeLabel = document.createElement('th');
-        homeLabel.innerHTML = this.translate('HOME');
+        homeLabel.innerHTML = this.nhlTranslate('HOME');
         homeLabel.setAttribute('colspan', 3);
         labelRow.appendChild(homeLabel);
 
@@ -173,7 +189,7 @@ Module.register('MMM-NHL', {
         labelRow.appendChild(vsLabel);
 
         const awayLabel = document.createElement('th');
-        awayLabel.innerHTML = this.translate('AWAY');
+        awayLabel.innerHTML = this.nhlTranslate('AWAY');
         awayLabel.setAttribute('colspan', 3);
         labelRow.appendChild(awayLabel);
 
@@ -187,47 +203,54 @@ Module.register('MMM-NHL', {
         const date = document.createElement('td');
         if (data.bsc === 'progress') {
             if (data.ts === 'PRE GAME') {
-                date.innerHTML = this.translate('PRE_GAME');
+                date.innerHTML = this.nhlTranslate('PRE_GAME');
                 date.classList.add('dimmed');
             } else if (['1st', '2nd', '3rd'].includes(data.ts.slice(-3))) {
                 const third = document.createElement('div');
-                third.innerHTML = this.translate(this.states[data.ts.slice(-3)]);
+                third.innerHTML = this.nhlTranslate(this.states[data.ts.slice(-3)]);
                 if (data.ts.slice(0, 3) !== 'END') {
                     third.classList.add('live');
                     date.appendChild(third);
                     const time = document.createElement('div');
                     time.classList.add('live');
-                    time.innerHTML = `${data.ts.slice(0, -4)} ${this.translate('TIME_LEFT')}`;
+                    time.innerHTML = `${data.ts.slice(0, -4)} ${this.nhlTranslate('TIME_LEFT')}`;
                     date.appendChild(time);
                 } else {
                     date.appendChild(third);
                 }
             }
+        } else if (data.bsc === '' && data.bs === 'PPD') {
+            date.innerHTML = this.nhlTranslate(this.states[data.bs]);
+            date.classList.add('dimmed');
         } else if (data.bsc === '' && Object.prototype.hasOwnProperty.call(data, 'starttime')) {
             date.innerHTML = moment(data.starttime).format(this.config.format);
         } else if (data.bsc === 'final') {
-            date.innerHTML = this.translate(this.states[data.bs]);
+            date.innerHTML = this.nhlTranslate(this.states[data.bs]);
             date.classList.add('dimmed');
         } else {
-            date.innerHTML = this.translate('UNKNOWN');
+            date.innerHTML = this.nhlTranslate('UNKNOWN');
             date.classList.add('dimmed');
         }
         row.appendChild(date);
 
         const homeTeam = document.createElement('td');
         homeTeam.classList.add('align-right');
-        const homeTeamSpan = document.createElement('span');
-        homeTeamSpan.innerHTML = this.teams[data.htv];
-        homeTeam.appendChild(homeTeamSpan);
+        if (this.config.showNames) {
+            const homeTeamSpan = document.createElement('span');
+            homeTeamSpan.innerHTML = this.teams[data.htv];
+            homeTeam.appendChild(homeTeamSpan);
+        }
         row.appendChild(homeTeam);
 
         const homeLogo = document.createElement('td');
-        const homeIcon = document.createElement('img');
-        homeIcon.src = this.file(`icons/${this.teams[data.htv]}.png`);
-        if (!this.config.colored) {
-            homeIcon.classList.add('icon');
+        if (this.config.showLogos) {
+            const homeIcon = document.createElement('img');
+            homeIcon.src = this.file(`icons/${this.teams[data.htv]}.png`);
+            if (!this.config.colored) {
+                homeIcon.classList.add('icon');
+            }
+            homeLogo.appendChild(homeIcon);
         }
-        homeLogo.appendChild(homeIcon);
         row.appendChild(homeLogo);
 
         const homeScore = document.createElement('td');
@@ -243,19 +266,23 @@ Module.register('MMM-NHL', {
         row.appendChild(awayScore);
 
         const awayLogo = document.createElement('td');
-        const awayIcon = document.createElement('img');
-        awayIcon.src = this.file(`icons/${this.teams[data.atv]}.png`);
-        if (!this.config.colored) {
-            awayIcon.classList.add('icon');
+        if (this.config.showLogos) {
+            const awayIcon = document.createElement('img');
+            awayIcon.src = this.file(`icons/${this.teams[data.atv]}.png`);
+            if (!this.config.colored) {
+                awayIcon.classList.add('icon');
+            }
+            awayLogo.appendChild(awayIcon);
         }
-        awayLogo.appendChild(awayIcon);
         row.appendChild(awayLogo);
 
         const awayTeam = document.createElement('td');
         awayTeam.classList.add('align-left');
-        const awayTeamSpan = document.createElement('span');
-        awayTeamSpan.innerHTML = this.teams[data.atv];
-        awayTeam.appendChild(awayTeamSpan);
+        if (this.config.showNames) {
+            const awayTeamSpan = document.createElement('span');
+            awayTeamSpan.innerHTML = this.teams[data.atv];
+            awayTeam.appendChild(awayTeamSpan);
+        }
         row.appendChild(awayTeam);
 
         appendTo.appendChild(row);