+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # MMM-NHL Changelog
 
+## [1.0.2]
+* Fixed Vegas Icon
+* Removed npm install from README (not needed and is huge)
+* Added options:
+     `gameTimeReloadInterval` for quicker reloads during games
+     `showNames` can toggle team names
+     `showLogos` can toggle team logos
+     `condensed` use shorter names (OT instead of Overtime)
+* Support PPD games (Postponed)
+
+
 ## [1.0.1]
 
 ### Added